# docs/analytics-demo/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QCross Instrument API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analytics/data-preview": {
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Data Preview",
        "operationId": "data_preview_api_v1_analytics_data_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Data Preview Api V1 Analytics Data Preview Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/frontier": {
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Frontier",
        "operationId": "frontier_api_v1_analytics_frontier_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/cross-validation": {
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Cross Validation",
        "operationId": "cross_validation_api_v1_analytics_cross_validation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/bootstrap": {
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Bootstrap",
        "operationId": "bootstrap_api_v1_analytics_bootstrap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/resampled-frontier": {
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Resampled Frontier",
        "operationId": "resampled_frontier_api_v1_analytics_resampled_frontier_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResampledFrontierRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/jobs/{job_id}": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Job Status",
        "operationId": "job_status_api_v1_analytics_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/jobs/{job_id}/result": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Job Result",
        "operationId": "job_result_api_v1_analytics_jobs__job_id__result_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyticsRequest": {
        "properties": {
          "instrument_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 15,
            "minItems": 2,
            "title": "Instrument Ids"
          },
          "currency": {
            "type": "string",
            "const": "USD",
            "title": "Currency",
            "default": "USD"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "covariance_method": {
            "type": "string",
            "enum": [
              "sample",
              "ledoit_wolf",
              "nonlinear",
              "cv_linear"
            ],
            "title": "Covariance Method",
            "default": "sample"
          },
          "shrinkage_target": {
            "type": "string",
            "enum": [
              "identity",
              "diagonal"
            ],
            "title": "Shrinkage Target",
            "default": "identity"
          },
          "risk_free_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Risk Free Rate",
            "default": 0.03
          },
          "objective": {
            "type": "string",
            "enum": [
              "gmv",
              "max_sharpe",
              "target_return"
            ],
            "title": "Objective",
            "default": "gmv"
          },
          "target_return": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 12.0,
                "minimum": -12.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Return"
          },
          "asset_constraints": {
            "additionalProperties": {
              "$ref": "#/components/schemas/WeightBounds"
            },
            "type": "object",
            "maxProperties": 15,
            "title": "Asset Constraints"
          },
          "group_constraints": {
            "additionalProperties": {
              "$ref": "#/components/schemas/WeightBounds"
            },
            "type": "object",
            "maxProperties": 30,
            "title": "Group Constraints"
          },
          "frontier_points": {
            "type": "integer",
            "maximum": 201.0,
            "minimum": 2.0,
            "title": "Frontier Points",
            "default": 51
          },
          "cv_folds": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 2.0,
            "title": "Cv Folds",
            "default": 5
          },
          "bootstrap_iterations": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Bootstrap Iterations",
            "default": 1000
          },
          "expected_block_length": {
            "type": "number",
            "maximum": 24.0,
            "minimum": 3.0,
            "title": "Expected Block Length",
            "default": 12
          },
          "random_seed": {
            "type": "integer",
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "title": "Random Seed",
            "default": 42
          },
          "bootstrap_objectives": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "gmv",
                    "max_sharpe"
                  ]
                },
                "type": "array",
                "maxItems": 2,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Bootstrap Objectives"
          },
          "risk_aversions": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array",
                "maxItems": 201,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Aversions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instrument_ids"
        ],
        "title": "AnalyticsRequest"
      },
...
      "FrontierResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "incomplete"
            ],
            "title": "Status"
          },
          "assets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assets"
          },
          "currency": {
            "type": "string",
            "const": "USD",
            "title": "Currency"
          },
          "covariance_method": {
            "type": "string",
            "enum": [
              "sample",
              "ledoit_wolf",
              "nonlinear",
              "cv_linear"
            ],
            "title": "Covariance Method"
          },
          "covariance_parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Covariance Parameters"
          },
          "covariance_diagnostics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Covariance Diagnostics"
          },
          "risk_free_rate": {
            "type": "number",
            "title": "Risk Free Rate"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "execution_seconds": {
            "type": "number",
            "title": "Execution Seconds"
          },
          "period": {
            "additionalProperties": true,
            "type": "object",
            "title": "Period"
          },
          "annual_expected_returns": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Annual Expected Returns"
          },
          "annual_covariance": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Annual Covariance"
          },
          "correlation": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "type": "array",
            "title": "Correlation"
          },
          "reproducibility": {
            "additionalProperties": true,
            "type": "object",
            "title": "Reproducibility"
          },
          "gmv": {
            "$ref": "#/components/schemas/PortfolioRead"
          },
          "max_sharpe": {
            "$ref": "#/components/schemas/PortfolioRead"
          },
          "maximum_return": {
            "$ref": "#/components/schemas/PortfolioRead"
          },
          "selected_portfolio": {
            "$ref": "#/components/schemas/PortfolioRead"
          },
          "equal_weight": {
            "$ref": "#/components/schemas/PortfolioRead"
          },
          "frontier": {
            "items": {
              "$ref": "#/components/schemas/PortfolioRead"
            },
            "type": "array",
            "title": "Frontier"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "status",
          "assets",
          "currency",
          "covariance_method",
          "covariance_parameters",
          "covariance_diagnostics",
          "risk_free_rate",
          "warnings",
          "execution_seconds",
          "period",
          "annual_expected_returns",
          "annual_covariance",
          "correlation",
          "reproducibility",
          "gmv",
          "max_sharpe",
          "maximum_return",
          "selected_portfolio",
          "equal_weight",
          "frontier"
        ],
        "title": "FrontierResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobRead": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "frontier",
              "cross-validation",
              "bootstrap",
              "resampled-frontier"
            ],
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "failed",
              "timed_out"
            ],
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "title": "Submitted At"
          },
          "elapsed_seconds": {
            "type": "number",
            "title": "Elapsed Seconds"
          },
          "timeout_seconds": {
            "type": "number",
            "title": "Timeout Seconds"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "result_url": {
            "type": "string",
            "title": "Result Url"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "kind",
          "status",
          "submitted_at",
          "elapsed_seconds",
          "timeout_seconds",
          "status_url",
          "result_url"
        ],
        "title": "JobRead"
      },
      "JobResultRead": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "frontier",
              "cross-validation",
              "bootstrap",
              "resampled-frontier"
            ],
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "failed",
              "timed_out"
            ],
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "title": "Submitted At"
          },
          "elapsed_seconds": {
            "type": "number",
            "title": "Elapsed Seconds"
          },
          "timeout_seconds": {
            "type": "number",
            "title": "Timeout Seconds"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "result_url": {
            "type": "string",
            "title": "Result Url"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FrontierResult"
              },
              {
                "$ref": "#/components/schemas/CrossValidationResult"
              },
              {
                "$ref": "#/components/schemas/BootstrapResult"
              },
              {
                "$ref": "#/components/schemas/ResampledFrontierResult"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "kind",
          "status",
          "submitted_at",
          "elapsed_seconds",
          "timeout_seconds",
          "status_url",
          "result_url",
          "result"
        ],
        "title": "JobResultRead"
      },
...
      "PortfolioRead": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PortfolioMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "diagnostics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Diagnostics"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "status",
          "weights",
          "metrics",
          "diagnostics"
        ],
        "title": "PortfolioRead"
      },
      "ResampledFrontierRequest": {
        "properties": {
          "instrument_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 15,
            "minItems": 2,
            "title": "Instrument Ids"
          },
          "currency": {
            "type": "string",
            "const": "USD",
            "title": "Currency",
            "default": "USD"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "covariance_method": {
            "type": "string",
            "enum": [
              "sample",
              "ledoit_wolf",
              "nonlinear",
              "cv_linear"
            ],
            "title": "Covariance Method",
            "default": "sample"
          },
          "shrinkage_target": {
            "type": "string",
            "enum": [
              "identity",
              "diagonal"
            ],
            "title": "Shrinkage Target",
            "default": "identity"
          },
          "risk_free_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Risk Free Rate",
            "default": 0.03
          },
          "objective": {
            "type": "string",
            "enum": [
              "gmv",
              "max_sharpe",
              "target_return"
            ],
            "title": "Objective",
            "default": "gmv"
          },
          "target_return": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 12.0,
                "minimum": -12.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Return"
          },
          "asset_constraints": {
            "additionalProperties": {
              "$ref": "#/components/schemas/WeightBounds"
            },
            "type": "object",
            "maxProperties": 15,
            "title": "Asset Constraints"
          },
          "group_constraints": {
            "additionalProperties": {
              "$ref": "#/components/schemas/WeightBounds"
            },
            "type": "object",
            "maxProperties": 30,
            "title": "Group Constraints"
          },
          "frontier_points": {
            "type": "integer",
            "maximum": 201.0,
            "minimum": 2.0,
            "title": "Frontier Points",
            "default": 51
          },
          "cv_folds": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 2.0,
            "title": "Cv Folds",
            "default": 5
          },
          "bootstrap_iterations": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Bootstrap Iterations",
            "default": 250
          },
          "expected_block_length": {
            "type": "number",
            "maximum": 24.0,
            "minimum": 3.0,
            "title": "Expected Block Length",
            "default": 12
          },
          "random_seed": {
            "type": "integer",
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "title": "Random Seed",
            "default": 42
          },
          "bootstrap_objectives": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "gmv",
                    "max_sharpe"
                  ]
                },
                "type": "array",
                "maxItems": 2,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Bootstrap Objectives"
          },
          "risk_aversions": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array",
                "maxItems": 201,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Aversions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instrument_ids"
        ],
        "title": "ResampledFrontierRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeightBounds": {
        "properties": {
          "min": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min",
            "default": 0
          },
          "max": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WeightBounds"
      }
    }
  }
}